refactor: app 디렉토리 구조 정리
- 영상 모니터링 페이지를 app/page.tsx에서 app/monitoring/page.tsx로 이동
- 루트 페이지(/)에서 /dashboard로 리다이렉트 처리
- 사이드바 네비게이션 업데이트 (/ → /monitoring)
- 미사용 my-wards 디렉토리 및 관련 파일 제거
- 공유 아이콘을 components/icons/shared-icons.tsx로 이동
- video 컴포넌트들의 CSS 모듈 import 경로 수정

Closes #155
