+++ app/globals.css
@@ -8,7 +8,6 @@
 @import '../styles/components.css';
 @import '../styles/dashboard.css';
 @import '../styles/beneficiaries.css';
-@import '../styles/my-wards.css';
 @import '../styles/stats.css';
 @import '../styles/staff.css';
 @import '../styles/settings.css';
