--- styles/my-wards.css
@@ -1,237 +1,0 @@
-/* --- My-Wards Page Styles --- */
-.my-wards-container {
-    max-width: 1400px;
-    margin: 0 auto;
-}
-
-/* Stats Grid */
-.my-wards-stats-grid {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    gap: 16px;
-    margin-bottom: 24px;
-}
-
-@media (max-width: 1024px) {
-    .my-wards-stats-grid {
-        grid-template-columns: repeat(2, 1fr);
-    }
-}
-
-@media (max-width: 640px) {
-    .my-wards-stats-grid {
-        grid-template-columns: 1fr;
-    }
-}
-
-/* Stat Card Styles */
-.stat-card {
-    background-color: var(--color-panel);
-    border: 1px solid var(--color-border);
-    border-radius: 14px;
-    padding: 16px;
-    box-shadow: var(--shadow-lifted);
-    position: relative;
-    overflow: hidden;
-}
-
-.stat-card.highlight {
-    box-shadow: var(--shadow-danger-glow);
-}
-
-.stat-card-icon {
-    position: absolute;
-    right: 10px;
-    top: 10px;
-    opacity: 0.12;
-}
-
-.stat-card-label {
-    font-size: 12px;
-    font-weight: 700;
-    color: var(--color-text-muted);
-    margin-bottom: 4px;
-}
-
-.stat-card-value {
-    display: flex;
-    align-items: baseline;
-    gap: 6px;
-}
-
-.stat-card-number {
-    font-size: 32px;
-    font-weight: 800;
-    color: var(--color-primary-dark);
-}
-
-.stat-card-number.highlight {
-    color: var(--color-danger-main);
-}
-
-.stat-card-unit {
-    color: var(--color-text-soft);
-    font-size: 13px;
-}
-
-.stat-card-footer {
-    margin-top: 8px;
-    font-size: 12px;
-    color: var(--color-text-soft);
-    font-weight: 600;
-}
-
-/* Filter and Search */
-.my-wards-filter-bar {
-    display: flex;
-    align-items: center;
-    gap: 12px;
-    margin-bottom: 16px;
-    flex-wrap: wrap;
-}
-
-.my-wards-filter-group {
-    display: flex;
-    background-color: var(--color-accent-soft);
-    padding: 6px;
-    border-radius: 12px;
-    gap: 6px;
-}
-
-.my-wards-filter-button {
-    padding: 10px 14px;
-    border-radius: 10px;
-    border: none;
-    background-color: transparent;
-    color: var(--color-text-soft);
-    font-weight: 700;
-    font-size: 13px;
-    cursor: pointer;
-    transition: all 150ms;
-}
-
-.my-wards-filter-button.active {
-    background-color: white;
-    color: var(--color-primary-dark);
-    box-shadow: var(--shadow-sm);
-}
-
-/* Table Styles */
-.my-wards-table-container {
-    background-color: white;
-    border-radius: var(--radius-lg);
-    border: 1px solid var(--color-border);
-    box-shadow: var(--shadow-sm);
-    overflow: hidden;
-}
-
-.my-wards-table-header {
-    padding: 16px 24px;
-    border-bottom: 1px solid var(--color-border);
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.my-wards-table-title {
-    font-size: 14px;
-    font-weight: 700;
-    color: var(--color-primary-dark);
-}
-
-.my-wards-table-head {
-    padding: 16px 24px;
-    font-size: 16px;
-    font-weight: 900;
-    color: var(--color-primary-dark);
-    background-color: var(--color-bg);
-}
-
-.my-wards-table-cell {
-    padding: 16px 24px;
-    font-size: 16px;
-    font-weight: 500;
-    color: var(--color-primary-dark);
-}
-
-.my-wards-table-row {
-    border-bottom: 1px solid var(--color-border);
-    transition: background-color 150ms;
-}
-
-.my-wards-table-row:hover {
-    background-color: var(--color-bg-elevated-1);
-}
-
-/* Ward Name Cell */
-.my-wards-name-cell {
-    display: flex;
-    align-items: center;
-    gap: 12px;
-}
-
-.my-wards-avatar {
-    width: 40px;
-    height: 40px;
-    border-radius: 12px;
-    background-color: var(--color-primary-light);
-    color: var(--color-primary-dark);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-weight: 800;
-    font-size: 16px;
-}
-
-.my-wards-name {
-    font-weight: 700;
-    color: var(--color-primary-dark);
-}
-
-.my-wards-org {
-    font-size: 12px;
-    color: var(--color-text-muted);
-}
-
-/* Status Badge */
-.my-wards-status-badge {
-    display: inline-flex;
-    align-items: center;
-    gap: 6px;
-    padding: 4px 10px;
-    border-radius: 999px;
-    font-size: 12px;
-    font-weight: 700;
-}
-
-.my-wards-status-badge.linked {
-    background-color: var(--color-success-soft);
-    color: var(--color-success-dark);
-}
-
-.my-wards-status-badge.pending {
-    background-color: var(--color-warning-soft);
-    color: #c2410c;
-}
-
-/* Resend Button */
-.my-wards-resend-btn {
-    padding: 8px 14px;
-    border-radius: 10px;
-    border: 1px solid var(--color-border);
-    background-color: white;
-    color: var(--color-primary-dark);
-    font-weight: 700;
-    font-size: 13px;
-    cursor: pointer;
-    transition: all 150ms;
-}
-
-.my-wards-resend-btn:hover {
-    background-color: var(--color-bg);
-}
-
-.my-wards-resend-btn:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-}
